Sample pass using pattern
This is done only to have something to point to when asked how Pattern
and Meander compare. I chose it because it appears to be a significant
and well-designed library which uses Meander in a substantial way.

This is a very simple 1:1 translation/port, not tested in any
substantial way. This port took me a little under an hour.

The couple of sample I found in the test suite produce the same result
in both implementations:

```clojure
(rewrite '(namespace hello (function hello [a b] (if a (+ a 1)))))

(rewrite '(namespace test-package.test-class
            (init x)
            (function f [y]
              (return
                {1 "hahaha"}))))
```